[runtime] Fix compilation warnings; NFCI
Fix the Flang code base so that WITH_WERROR=on can be used at CMake time to
enable stricter linting. This patch mainly addresses the following types of
warnings:

- -Wcast-qual warnings as a result of mixing (char *) fields and parameters
  with (const char *) values, especially string literals. Most of these can
  be trivially fixed by changing the type of the fields/parameters in question
  to (const char *). Tricky cases are handled with strdup or explicit casts.

- -Wunused-variable and -Wsometimes-uninitialized warnings. Unused variables
  are deleted, and ones that are used conditionally are initialized with
  sensible defaults.

- -Wmissing-field-initializers warnings for partial initializers in data
  structure definitions.

- -Wformat warnings (e.g. "%d" vs. "%ld", "%p" for non-(void *) arguments).

+++ runtime/flang/ftncharsup.c
@@ -70,7 +70,7 @@ Ftn_str_copy(int n, char *to, int to_len, ...)
   if (to_len <= 0) {
     return;
   }
-  if (n <= (sizeof(aa) / sizeof(SRC_STR))) {
+  if (n <= (int)(sizeof(aa) / sizeof(SRC_STR))) {
     qq = src_p = aa;
     src_p_allocd = 0;
   } else {
@@ -305,7 +305,7 @@ Ftn_str_copy_klen(int n, char *to, int64_t to_len, ...)
   if (to_len <= 0) {
     return;
   }
-  if (n <= (sizeof(aa) / sizeof(SRC_STR))) {
+  if (n <= (int)(sizeof(aa) / sizeof(SRC_STR))) {
     qq = src_p = aa;
     src_p_allocd = 0;
   } else {
